# swagger.yaml
openapi: 3.0.0
info:
  title: Online Learning Management System API
  version: 1.0.0
  description: APIs for user registration and login in an online learning management system.
paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - username
                - email
                - password
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              example:
                data: User registered successfully
        '400':
          description: Bad Request - Invalid input data
          content:
            application/json:
              example:
                data: Invalid input data
        '409':
          description: Conflict - Username or email already exists
          content:
            application/json:
              example:
                data: Username or email already exists
  /login:
    post:
      summary: Log in as a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              example:
                message1: Login successful
                token: <JWT Token>
        '400':
          description: Bad Request - Invalid input data
          content:
            application/json:
              example:
                data: Invalid input data
        '401':
          description: Unauthorized - Username or password do not match
          content:
            application/json:
              example:
                data: Username or password do not match